Move update helpers to LockedManifest, factor out ensure_global_lockfile
